+++ ipc.c
@@ -26,9 +26,9 @@ static void
 ipc_send_log_internal(enum DataType type, char *format, va_list va_args) {
     Message *message;
     char buffer[8192];
-    int64 n;
-    int64 last_whitespace_index;
-    int64 current_column;
+    int32 n;
+    int32 last_whitespace_index;
+    int32 current_column;
 
     n = vsnprintf(buffer, SIZEOF(buffer), format, va_args);
 
@@ -39,7 +39,7 @@ ipc_send_log_internal(enum DataType type, char *format, va_list va_args) {
 
     last_whitespace_index = -1;
     current_column = 0;
-    for (int64 i = 0; i < n; i += 1) {
+    for (int32 i = 0; i < n; i += 1) {
         if ((buffer[i] == ' ') || (buffer[i] == '\t') || (buffer[i] == '\n')) {
             last_whitespace_index = i;
         }
@@ -131,21 +131,21 @@ ipc_send_tree(int32 side,
                   int64 size, int64 mtime) {
     // clang-format on
     Message *message;
-    int64 target_len;
-    int64 pattern_len;
+    int32 target_len;
+    int32 pattern_len;
 
     g_mutex_lock(&cecup.ui_arena_mutex);
     message = xarena_push(cecup.ui_arena, ALIGN16(SIZEOF(Message)));
     memset64(message, 0, SIZEOF(Message));
 
-    message->filepath_len = strlen64(path);
+    message->filepath_len = (int32)strlen64(path);
     g_mutex_lock(&cecup.row_arena_mutex);
     message->filepath
         = xarena_push(cecup.row_arena, ALIGN16(message->filepath_len + 1));
     memcpy64(message->filepath, path, message->filepath_len + 1);
 
     if (link_target) {
-        target_len = strlen64(link_target);
+        target_len = (int32)strlen64(link_target);
         message->link_target_len = target_len;
         message->link_target
             = xarena_push(cecup.row_arena, ALIGN16(target_len + 1));
@@ -153,7 +153,7 @@ ipc_send_tree(int32 side,
     }
 
     if (ignore_pattern) {
-        pattern_len = strlen64(ignore_pattern);
+        pattern_len = (int32)strlen64(ignore_pattern);
         message->ignore_pattern_len = pattern_len;
         message->ignore_pattern
             = xarena_push(cecup.row_arena, ALIGN16(pattern_len + 1));
